+++ ChangeLog.md
@@ -16,6 +16,8 @@
 
 * Fixed internal error with assertions with intra-tensor dependencies e.g. `forall x . x ! 0 < x ! 1`.
 
+* The branches of an `if` may now have different gradients, e.g. `if c then x else 0.0`.
+
 ### Solver backend
 
 * Fixed internal error that occassionally happened with non-linear specifications.